Move all test helper classes out of AbstractTransportTest so they can be used elsewhere (#12125)
