+++ inprocess/src/test/java/io/grpc/inprocess/InProcessTransportTest.java
@@ -36,10 +36,14 @@ import io.grpc.Status.Code;
 import io.grpc.StatusRuntimeException;
 import io.grpc.internal.AbstractTransportTest;
 import io.grpc.internal.ClientStream;
+import io.grpc.internal.ClientStreamListenerBase;
 import io.grpc.internal.GrpcUtil;
 import io.grpc.internal.InternalServer;
 import io.grpc.internal.ManagedClientTransport;
+import io.grpc.internal.MockServerTransportListener;
+import io.grpc.internal.MockServerTransportListener.StreamCreation;
 import io.grpc.internal.ServerStream;
+import io.grpc.internal.ServerStreamListenerBase;
 import io.grpc.internal.testing.TestStreamTracer;
 import io.grpc.stub.ClientCalls;
 import io.grpc.testing.GrpcCleanupRule;
